requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
pydantic-settings
psutil

# ML / AI
torch
torchaudio
transformers
accelerate
bitsandbytes
Pillow
nemo_toolkit[asr]
sentence-transformers
opencv-python  # For video processing (SAM 3)
sounddevice
soundfile

# IoT backends
aiomqtt
httpx
websockets

# Telephony (external communications)
signalwire>=2.0.0
# twilio  # Alternative provider

# Discovery
zeroconf>=0.80.0

# Database
asyncpg>=0.29.0

# Natural language parsing
dateparser>=1.2.0

# Orchestration / Voice pipeline
webrtcvad
piper-tts
kokoro>=0.7.0

# Optional: For GGUF model support
# llama-cpp-python

# Optional: Flash Attention 2 for faster inference (Ampere+ GPUs)
# Install separately: pip install flash-attn --no-build-isolation
rich>=13.0.0

# LangGraph for agent orchestration
langgraph>=0.2.0
langchain-core>=0.3.0

# Autonomous task scheduling
apscheduler>=3.10,<4.0

# Email draft generation (Anthropic Claude)
anthropic>=0.39.0

# MCP (Model Context Protocol) — CRM and Email MCP servers
mcp>=1.26.0

# External data producers (news feeds)
feedparser>=6.0
trafilatura>=1.6

# B2B review scraping (anti-detection HTTP client + HTML parsing)
curl_cffi>=0.7.0
beautifulsoup4>=4.12.0
